Extract generic JWT token builder code
In order to use JWT tokens for something else than download tokens,
move the download token specific code into a DownloadTokenBuilder.
This way adding support for WebSockify tokens would only mean
implementing that new token specific code.
