+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/activationkey/ActivationKeyHandler.java
@@ -54,7 +54,8 @@ import com.redhat.rhn.manager.token.ActivationKeyManager;
 
 import com.suse.manager.utils.MachinePasswordUtils;
 import com.suse.manager.webui.services.SaltStateGeneratorService;
-import com.suse.manager.webui.utils.TokenBuilder;
+import com.suse.manager.webui.utils.DownloadTokenBuilder;
+
 import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
 import org.hibernate.NonUniqueObjectException;
@@ -262,7 +263,7 @@ public class ActivationKeyHandler extends BaseHandler {
             throw new AuthenticationException("wrong machine password.");
         }
 
-        TokenBuilder tokenBuilder = new TokenBuilder(minion.getOrg().getId());
+        DownloadTokenBuilder tokenBuilder = new DownloadTokenBuilder(minion.getOrg().getId());
         tokenBuilder.useServerSecret();
         tokenBuilder.setExpirationTimeMinutesInTheFuture(
                 Config.get().getInt(
